subfolder breakdown and renames
