+++ .eslintrc.json
@@ -17,6 +17,15 @@
         "sourceType": "module"
     },
     "rules": {
-        "@typescript-eslint/explicit-module-boundary-types": "off"
+        "@typescript-eslint/explicit-module-boundary-types": "off",
+        "@typescript-eslint/naming-convention": [
+            "warn",
+            {
+                "selector": "enum",
+                "format": [
+                    "StrictPascalCase"
+                ]
+            }
+        ]
     }
 }
