+++ src/index.ts
@@ -1,17 +1,3 @@
-export { Trace, ITraceContext, TraceContext, useTrace } from './analytics/Trace.js'
-export { TraceEvent } from './analytics/TraceEvent.js'
-
-export {
-  BROWSER,
-  CUSTOM_USER_PROPERTIES,
-  ElementName,
-  Event,
-  EventName,
-  ModalName,
-  PageName,
-  SWAP_PRICE_UPDATE_USER_RESPONSE,
-  SectionName,
-  WALLET_CONNECTION_RESULT,
-} from './analytics/constants.js'
-
-export { initializeAnalytics, sendAnalyticsEvent, user } from './analytics/index.js'
+export * from './analytics/Trace.js'
+export * from './analytics/TraceEvent.js'
+export * from './analytics/index.js'
